usage tests: migrated all tests
